Finish model.feed & its tests
