+++ tests/test_model.py
@@ -1,14 +1,23 @@
 import unittest
 import os
 import sys
+
+from model.feed import Feed
+
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
 from datetime import timedelta, date, datetime
 from model.model import *
-from unittest.mock import Mock, patch
 
 
-class ModelTestCase(unittest.TestCase):
+class ArticleTestCase(unittest.TestCase):
+
+    def test_article(self):
+        # TODO: Create test for Article
+        pass
+
+
+class FeedTestCase(unittest.TestCase):
 
     def test_feed_add_new(self):
         feed_name = "Feed Name"
@@ -25,7 +34,7 @@ class ModelTestCase(unittest.TestCase):
         article_4 = Article("Article 4", "Link 4", (datetime.now() - timedelta(days=4)))    # 4 days ago
         feed.update([article_2, article_3, article_4])
         feed.add_new(article_1)
-        self.assertEqual(feed.current_article, article_1) # Should default to the new article
+        self.assertEqual(feed.current_article, article_1) # Should default to the new article.py
 
 
     def test_feed_is_empty(self):
@@ -65,7 +74,7 @@ class ModelTestCase(unittest.TestCase):
 
         self.assertTrue(feed.is_sorted())
 
-    def test_feed_get_next(self):
+    def test_feed_move_to_next(self):
         feed_name = "Feed Name"
         feed = Feed(feed_name)
 
@@ -81,11 +90,9 @@ class ModelTestCase(unittest.TestCase):
         article_4 = Article("Article 4", "Link 4", (datetime.now() - timedelta(days=4)))    # 4 days ago
         feed.update([article_2, article_3, article_4])
 
-        self.assertTrue(feed.move_to_next())
         self.assertEqual(feed.current_article, article_2)
 
-        feed.move_to_next()
-
+        self.assertTrue(feed.move_to_next())
         self.assertEqual(feed.current_article, article_3)
 
         feed.move_to_next()
@@ -94,7 +101,7 @@ class ModelTestCase(unittest.TestCase):
 
         feed.move_to_next()
 
-        self.assertEqual(feed.current_article, article_1) # Should loop around end to start
+        self.assertEqual(feed.current_article, article_2) # Should loop around end to start
 
     def test_feed_update(self):
         article_1 = Article("Article 1", "Link 1", (datetime.now() - timedelta(days=1)))    # 1 day ago (most recent)
@@ -119,6 +126,9 @@ class ModelTestCase(unittest.TestCase):
         feed.update([article_1, article_4, article_2])
         self.assertEqual(feed.current_article, article_4) # Current should stay the same between updates if possible.
 
+
+class ModelTestCase(unittest.TestCase):
+
     def test_model(self):
         # TODO: Create test for Model
         pass
@@ -128,7 +138,7 @@ class ModelTestCase(unittest.TestCase):
         pass
 
     def test_model_add_article(self):
-        # TODO: Create test for Model.add(article)
+        # TODO: Create test for Model.add(article.py)
         pass
 
     def test_model_remove(self):
@@ -138,7 +148,6 @@ class ModelTestCase(unittest.TestCase):
     def test_parse(self):
         # TODO: Create test for parse()
         pass
-
 
 
 if __name__ == '__main__':
